chore: cleanup last context menu
